Allow tests/dockerfiles/build.py to be invoked by users
The `tests/dockerfiles/build.py` script allows a "base" Docker image
to be built for testing, which includes basic prerequisites for
pyinfra (such as openssh), but does not install Dokku.

The script was intended just to be invoked as part of GitHub CI builds,
but users may find it helpful. So we adjust the logic so that it
can be run by a user from the command line, without creating files
the script would expect in a CI environment.